chore: #33 Candidate 관련 코드 폴더 경로 변경
