+++ src/main/java/com/otechdong/moyeo/domain/candidatePlace/service/CandidatePlaceService.java
@@ -1,0 +1,11 @@
+package com.otechdong.moyeo.domain.candidatePlace.service;
+
+import com.otechdong.moyeo.domain.meeting.dto.MeetingResponse;
+import com.otechdong.moyeo.domain.member.entity.Member;
+
+public interface CandidatePlaceService {
+    MeetingResponse.MeetingAddCandidatePlace addCandidatePlace(
+            Member member,
+            Long meetingId,
+            Long placeId);
+}
